+++ dev/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @khanacademy/perseus-dev-ui
 
+## 5.6.6
+
+### Patch Changes
+
+-   Updated dependencies [[`aab0bf4b8b`](https://github.com/Khan/perseus/commit/aab0bf4b8bcb10347f24854403ee27581212df5c), [`9a5605b82c`](https://github.com/Khan/perseus/commit/9a5605b82ca6038f326bf70c6d0959bb2227b754)]:
+    -   @khanacademy/perseus-core@18.8.0
+    -   @khanacademy/kmath@2.1.5
+    -   @khanacademy/math-input@26.1.6
+    -   @khanacademy/perseus-linter@4.2.5
+
+## 5.6.5
+
+### Patch Changes
+
+-   Updated dependencies [[`52b2b3ae54`](https://github.com/Khan/perseus/commit/52b2b3ae54953ff6969e8962858d46d025661be8), [`62029ee700`](https://github.com/Khan/perseus/commit/62029ee700db45a646a0addf0e90ff4df73aef3a)]:
+    -   @khanacademy/perseus-core@18.7.0
+    -   @khanacademy/kmath@2.1.4
+    -   @khanacademy/math-input@26.1.5
+    -   @khanacademy/perseus-linter@4.2.4
+
 ## 5.6.4
 
 ### Patch Changes